add system-components
