+++ package.json
@@ -7,13 +7,14 @@
     "gatsby": "^1.9.247",
     "gatsby-link": "^1.6.40",
     "gatsby-plugin-react-helmet": "^2.0.10",
-    "gatsby-plugin-styled-components": "^2.0.11",
     "gatsby-plugin-robots-txt": "^1.0.2",
+    "gatsby-plugin-styled-components": "^2.0.11",
     "react-helmet": "^5.2.0",
     "react-powerplug": "^1.0.0-alpha.7",
     "sanitize.css": "^5.0.0",
     "styled-components": "^3.3.2",
-    "styled-system": "^2.2.5"
+    "styled-system": "^2.2.5",
+    "system-components": "^2.0.3"
   },
   "keywords": [
     "gatsby"
